got all the crud stuff done
